# specs/main/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Inspiration Box API
  version: 0.1.0
servers:
  - url: http://localhost:8000
paths:
  /memos:
    get:
      summary: List memos
      parameters:
        - in: query
          name: tag
          schema:
            type: string
          description: Filter memos by tag name
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: A list of memos
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Memo'
                  total:
                    type: integer
    post:
      summary: Create a memo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoCreate'
      responses:
        '201':
          description: Memo created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memo'
        '400':
          description: Validation error
  /memos/{id}:
    get:
      summary: Get a memo by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Memo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memo'
        '404':
          description: Not found
components:
  schemas:
    Memo:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - content
        - tags
        - created_at
    MemoCreate:
      type: object
      properties:
        content:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - content